requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# AI and ML
google-generativeai>=0.3.0

# Vector databases
qdrant-client>=1.7.0

# Document processing
pypdf>=3.17.0
python-docx>=1.1.0
pandas>=2.1.0
openpyxl>=3.1.0
Pillow>=10.1.0
opencv-python>=4.8.0
pytesseract>=0.3.10
PyMuPDF>=1.23.0

# Database and storage
cassandra-driver>=3.28.0
redis>=5.0.0
elasticsearch>=8.11.0
minio>=7.2.0

# Message queuing
kafka-python>=2.0.0
confluent-kafka>=2.3.0

# Monitoring and observability
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
jaeger-client>=4.8.0

# Utilities
python-multipart>=0.0.6
aiofiles>=23.2.0
httpx>=0.25.0
tenacity>=8.2.0
structlog>=23.2.0
rich>=13.7.0

# Security
cryptography>=41.0.0
passlib[bcrypt]>=1.7.0
python-jose[cryptography]>=3.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0

# Data Processing
polars==0.19.19
pyarrow==14.0.2

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pillow==10.1.0
opencv-python==4.8.1.78

# Vector Databases
qdrant-client==1.7.0

# Distributed Storage
cassandra-driver==3.28.0
redis==5.0.1
minio==7.2.0
elasticsearch==8.11.1

# Message Queue
kafka-python==2.0.2
confluent-kafka==2.3.0

# Monitoring and Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
jaeger-client==4.8.0

# API and Communication
grpcio==1.59.3
grpcio-tools==1.59.3
httpx==0.25.2
websockets==12.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development and Utilities
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Configuration and Environment
pyyaml==6.0.1
toml==0.10.2

# Logging
structlog==23.2.0
rich==13.7.0

# Async Support
asyncio-mqtt==0.16.1
aiofiles==23.2.1
aiohttp==3.9.1

# Data Validation
marshmallow==3.20.1
cerberus==1.3.5

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78
imageio==2.33.1

# Text Processing
nltk==3.8.1
spacy==3.7.2
textblob==0.17.1

# Database ORM
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Caching
diskcache==5.6.3
cachetools==5.3.2

# Serialization
msgpack==1.0.7
orjson==3.9.10

# Networking
aiohttp==3.9.1
requests==2.31.0

# AI and ML
pytesseract==0.3.10
PyMuPDF==1.23.8

# Utilities
tenacity==8.2.3 